% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ccs_tree.R
\name{ccs_tree}
\alias{ccs_tree}
\title{\code{ccs_tree} produces tree of Clinical Classification Software (CCS) codes
based on root node.}
\usage{
ccs_tree(group = NULL)
}
\arguments{
\item{group}{character string specifying that only codes beginning with group 
will be included in tree The default is NULL (returns full tree of CCS codes)}
}
\value{
A list containing the following elements:
\describe{
\item{tr}{directed igraph object; the tree of CCS codes.}
\item{ccs_icd_mapping}{data frame specifying the ICD9 codes corresponding to each 
CCS code. Useful for converting from ICD9 to CCS.}
}
}
\description{
\code{ccs_tree} produces tree of Clinical Classification Software (CCS) codes
based on root node.
}
\details{
Adapted from \code{icd9_map_multi_ccs} in the \code{icd} package.
}
\examples{
tr <- ccs_tree("7.4")$tr
# Plot tree
library(ggtree)
library(ggplot2)
ggtree(tr, ladderize = F, layout = "slanted") + 
geom_tiplab(geom = "label") + 
geom_nodelab(geom = "label") +
theme(plot.margin = unit(c(0, 1.5, 0, 0.2), "cm")) +
coord_cartesian(clip = "off") + 
scale_y_reverse()
}
\concept{Tree functions}
