% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moretrees.R, R/moretrees_wrapper.R
\docType{package}
\name{moretrees}
\alias{moretrees}
\alias{moretrees-package}
\title{moretrees: A package for fitting Multi-Outcome Regression with 
Tree-structured Shrinkage (MOReTreeS)}
\usage{
moretrees(X, W = NULL, y, outcomes, tr, random_init = FALSE,
  initial_values = NULL, family = "bernoulli", ci_level = 0.95,
  get_ml = FALSE, update_hyper_freq = 50,
  print_freq = update_hyper_freq, hyper_fixed = NULL, tol = 1e-08,
  tol_hyper = 1e-06, max_iter = 5000, nrestarts = 3,
  keep_restarts = nrestarts > 1, parallel = nrestarts > 1,
  log_restarts = nrestarts > 1, log_dir = getwd(),
  random_init_vals = list(omega_lims = c(0.5, 1.5), tau_lims = c(0.5,
  1.5), sigma2_lims = c(0.5, 1.5), eta_sd_frac = 0.2, mu_sd_frac = 0.2,
  delta_sd_frac = 0.2))
}
\arguments{
\item{X}{An n x K matrix of exposure data, where K is the dimension of the exposure.
Grouping of the outcomes will be based on their relationships with the variables in X.}

\item{W}{Matrix of covariates of dimension n x m.
Coefficients for these variables do not affect grouping of the outcomes.}

\item{y}{Vector of length n containing outcomes data.
If family = "bernoulli", y must be an integer vector where 1 = success, 0 = failure.
If family = "gaussian", y must be a numeric vector containing continuous data.}

\item{outcomes}{Character vector of length n. outcomes[i] is a string indicating the 
outcome experienced by unit i.}

\item{tr}{A directed igraph object. This is a tree representing the relationships
among the outcomes. The leaves represent individual outcomes, and internal nodes
represent outcome categories consisting of their leaf descendants. All nodes
of tr must have unique names as given by names(V(tr)). The names of the leaves must 
be equal to the unique elements of outcomes. The vertices of tr, V(tr), may have 
an attribute 'levels' containing integer values from 1 to max(V(tr)$levels). 
In this case, the levels attribute specifies groups of nodes that share common 
hyperparameters rho[f], tau[f], and omega[f]. If V(tr)$levels is NULL, 
levels will be automatically chosen based on distance from the root node of the tree.}

\item{random_init}{If TRUE, initial values will be randomly permuted.}

\item{family}{A string specifying the distribution of the outcomes: 
either "bernoulli" (for classification) or "gaussian" (for regression)}

\item{ci_level}{A number between 0 and 1 giving the desired credible interval. 
For example, ci_level = 0.95 (the default) returns a 95\% credible interval}

\item{get_ml}{If TRUE, moretrees will also return the maximum likelihood estimates of the
coefficients for each outcome group discovered by the model. The default is FALSE.}

\item{update_hyper_freq}{How frequently to update hyperparameters. 
Default = every 50 iterations.}

\item{print_freq}{How often to print out iteration number and current value of epsilon
(the difference in objective function value for the two most recent iterations).}

\item{hyper_fixed}{Fixed values of hyperprior parameters for rho.
If family = "bernoulli", this should be a list including the following elements:
a_rho, b_rho (parameters of beta prior on rho for each level)
If family = "gaussian", in addition to the above, the list should also include:
sigma2 (variance of residuals)}

\item{tol}{Convergence tolerance for ELBO. Default = 1E-8.}

\item{tol_hyper}{If hyper_method = "EB", the convergence tolerance for ELBO
between subsequent hyperparmeter updates. Typically a more generous
tolerance than tol. Default = 1E-6.}

\item{nrestarts}{Number of random re-starts of the VI algorithm. The result that 
gives the highest ELBO will be returned. It is recommended to choose nrestarts > 1.
The default is 3.}

\item{keep_restarts}{If TRUE, the results from all random restarts will be returned.
If FALSE, only the restart with the highest ELBO is returned.}

\item{parallel}{If TRUE, the random restarts will be run in parallel. It is recommended
to first set the number of cores using doParallel::registerDoParallel(). Otherwise,
the default number of cores specified by the doParallel package will be used.}

\item{log_restarts}{If TRUE, progress of each random restart will be logged to a text
file in log_dir.}

\item{log_dir}{Directory for logging progress of random restarts.
Default is the working directory.}

\item{random_init_vals}{If random_init = TRUE, 
this is a list containing parameters for randomly permuting the inital values.
The list contains the following elements:}

\item{maxiter}{Maximum number of iterations of the VI algorithm.}
}
\value{
A list containing the following elements:
1. estimated coefficients and credible intervals; 
2. outputs from variational inference algorithm
}
\description{
Fit MOReTreeS to: normally distributed outcome data (moretrees_normal) or
binary data (moretrees_logistic).

All the details go here!
}
\section{Here are some functions}{

All about functions!
}

\section{Model Description}{

Describe MOReTreeS model and all parameters here.
}

\seealso{
Other MOReTreeS functions: \code{\link{moretrees_compute_betas}},
  \code{\link{moretrees_compute_thetas}},
  \code{\link{moretrees_design_tree}},
  \code{\link{moretrees_init_logistic}}
}
\concept{MOReTreeS functions}
