% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moretrees_compute_betas.R
\name{moretrees_compute_betas}
\alias{moretrees_compute_betas}
\title{\code{moretrees_compute_betas} Computes MOReTreeS estimates and credible
  intervals for exposure effects from model output}
\usage{
moretrees_compute_betas(mod, ci_level, A_leaves, outcomes)
}
\arguments{
\item{mod}{List containing outputs from spike and slab VI algorithm}

\item{ci_level}{A number between 0 and 1 giving the desired credible interval.
For example, ci_level = 0.95 (the default) returns a 95\% credible interval.}

\item{A_leaves}{pL x p sparse ancestor Matrix where rows correspond to leaves
of tree (outcomes) and columns correspond to nodes on tree. Results in mod
must have same ordering as columns of A_leaves.}
}
\value{
A list containing the following elements:
1. beta_moretrees: estimated coefficients and credible intervals by group;
2. beta_est: estimated coefficients and credible intervals by outcome;
}
\description{
\code{moretrees_compute_betas} Computes MOReTreeS estimates and credible
  intervals for exposure effects from model output
}
\seealso{
Other Processing model output: \code{\link{ml_by_group}},
  \code{\link{moretrees_compute_thetas}}
}
\concept{Processing model output}
