% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moretrees_design_tree.R
\name{moretrees_design_tree}
\alias{moretrees_design_tree}
\title{\code{moretrees_design_tree} converts outcome, exposure, and covariate data
  into format suitable for analysis using MOReTreeS.}
\usage{
moretrees_design_tree(y, X, W = NULL, outcomes, tr)
}
\arguments{
\item{y}{Vector of length n containing outcomes data.
If family = "bernoulli", y must be an integer vector where 1 = success, 0 = failure.
If family = "gaussian", y must be a numeric vector containing continuous data.}

\item{X}{An n x K matrix of exposure data, where K is the dimension of the exposure.
Grouping of the outcomes will be based on their relationships with the variables in X.}

\item{W}{Matrix of covariates of dimension n x m.
Coefficients for these variables do not affect grouping of the outcomes.
Default is NULL (no covariates).}

\item{outcomes}{is a character vector of length n, where entry i}

\item{tr}{is an igraph tree, where the leaves represent outcomes.
The vertices of tr, V(tr), may have an attribute 'levels' containing integer
values from 1 to max(V(tr)$levels). In this case, the levels attribute specifies
groups of nodes that share common hyperparameters rho[f], tau[f], and omega[f].
If V(tr)$levels is NULL, moretrees_design_tree will specify levels based on
distance from the root node of the tree.}
}
\value{
A list containing the following elements:
y: Re-ordered outcome vector.
X: Re-ordered exposure matrix.
W: Re-ordered covariate matrix.
A: Matrix for transforming node-level parameters to outcome-level parameters.
outcomes_units: list of length equal to the number of unique outcomes. Each element of
the list is an integer vector indicating which units (entries of y_reord, rows of X_reord)
correspond to each outcomes.
outcomes_nodes: list of length equal to the number of unique nodes. Each element of
the list is an integer vector indicating which outcomes are descendants of each node.
ancestors: list of length equal to the number of unique outcomes. Each element of the 
list is an integer vector indicating which nodes on the tree (including leaves) are ancestors
of the corresponding outcome.
levels: an integer vector of length equal to the number of unique nodes. Contains
integer values from 1 to max(levels), where the values indicate groups of nodes
that share common hyperparameters rho[f], tau[f], and omega[f], where f is the value
in levels.
}
\description{
\code{moretrees_design_tree} converts outcome, exposure, and covariate data
  into format suitable for analysis using MOReTreeS.
}
\concept{Data processing functions}
