% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moretrees_normal.R
\name{moretrees_normal}
\alias{moretrees_normal}
\title{ssMOReTreeS with a Gaussian outcome,
with model fitting via variational inference.}
\usage{
moretrees_normal(y, X, tol = 1e-04, update_hyper = T,
  update_hyper_freq = 10, print_freq = 10)
}
\arguments{
\item{y}{Vector of outcomes data.}

\item{X}{Array of design matrices for each variable group.}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
ssMOReTreeS with a Gaussian outcome,
with model fitting via variational inference.
}
\examples{
add(1, 1)
}
