% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.moretrees_result.R
\name{print.moretrees_result}
\alias{print.moretrees_result}
\title{\code{print.moretrees_result} summarizes the results from \code{moretrees()}.}
\usage{
\method{print}{moretrees_result}(x, ...)
}
\arguments{
\item{x}{Output from \code{moretrees()}.}

\item{...}{Arguments passed to summary and printing methods.}
}
\value{
Summary showing, for each outcome group discovered by MOReTreeS,
the exposure coefficients, 95\% credible intervals, number of outcomes
per group, and number of matched pairs.
}
\description{
\code{print.moretrees_result} summarizes the results from \code{moretrees()}.
}
\examples{
# See vignette
vignette("moretrees")
}
\seealso{
Other MOReTrees results: \code{\link{plot.moretrees_result}},
  \code{\link{print.summary.moretrees_compact}},
  \code{\link{print.summary.moretrees_long}},
  \code{\link{summary.moretrees_result}}
}
\concept{MOReTrees results}
