% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.summary.moretrees_long.R
\name{print.summary.moretrees_long}
\alias{print.summary.moretrees_long}
\title{Long-form printing of MOReTreeS model fits}
\usage{
\method{print}{summary.moretrees_long}(x, print_outcomes = T,
  digits = max(3L, getOption("digits") - 3L), ...)
}
\arguments{
\item{x}{output from \code{summary.moretrees_result} with \code{compact = FALSE}.}

\item{print_outcomes}{If \code{TRUE}, for each discovered group the full list
of outcomes will be printed. Set this to \code{FALSE} if these outcome lists
make output difficult to read.}

\item{digits}{Number of significant digits to print.}

\item{...}{Not used.}
}
\value{
Summary showing, for each outcome group discovered by MOReTreeS,
the exposure coefficients, 95\% credible intervals, number of outcomes
per group, and number of matched pairs.
}
\description{
\code{print.summary.moretrees_long} is a print method for class 
\code{summary.moretrees_long}.
}
\examples{
# See vignette
vignette("moretrees")
}
\seealso{
Other MOReTrees results: \code{\link{plot.moretrees_result}},
  \code{\link{print.moretrees_result}},
  \code{\link{print.summary.moretrees_compact}},
  \code{\link{summary.moretrees_result}}
}
\concept{MOReTrees results}
