% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike_and_slab_normal.R
\name{spike_and_slab_normal}
\alias{spike_and_slab_normal}
\title{Group spike and slab variable selection with Gaussian outcome}
\usage{
spike_and_slab_normal(y, X, tol = 1e-04, update_hyper = T,
  update_hyper_freq = 10, print_freq = 10)
}
\arguments{
\item{y}{Vector of outcomes data.}

\item{X}{Array of design matrices for each variable group.}

\item{tol}{Convergence tolerance for ELBO. Default = 1E-4.}

\item{update_hyper}{Update hyperparameters? Default = TRUE.}

\item{update_hyper_freq}{How frequently to update hyperparameters. Default = every 10 iterations.}

\item{print_freq}{How often to print out iteration number. Default = every 10 iterations.}
}
\value{
A list of variational parameters.
}
\description{
Here's a brief description.
\code{spike_and_slab_normal} performs group variable selection via a spike
and slab prior. The posterior is approximated via variational inference.
This function returns the parameters of the variational approximation.
}
\details{
All the details go here!
}
\section{Model Description}{

Describe group spike and slab prior and all parameters here.
}

\examples{
Add this later from test file.
}
\concept{spike and slab functions}
