% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.moretrees_result.R
\name{summary.moretrees_result}
\alias{summary.moretrees_result}
\title{\code{summary.moretrees_result} summarizes the results from \code{moretrees()}.}
\usage{
\method{summary}{moretrees_result}(object, transform = "exp",
  coeff_type = "moretrees", compact = F, ...)
}
\arguments{
\item{object}{Output from \code{moretrees()}. 
An object of class "moretrees_result".}

\item{transform}{Either "exp", "identity", or "er" (excess rate)}

\item{coeff_type}{Either "moretrees" or "clr"}

\item{compact}{If \code{TRUE}, a more compact summary of results is printed.
Only works well when the dimension of the exposure variable is low
(say, < 4); otherwise the table takes up too much horizontal space.
Default is \code{FALSE}.}

\item{...}{Not used.}
}
\value{
Summary showing, for each outcome group discovered by MOReTreeS,
the exposure coefficients, 95\% credible intervals, number of outcomes
per group, and number of matched pairs.
}
\description{
\code{summary.moretrees_result} summarizes the results from \code{moretrees()}.
}
\examples{
# See vignette
vignette("moretrees")
}
\seealso{
Other MOReTrees results: \code{\link{plot.moretrees_result}},
  \code{\link{print.moretrees_result}},
  \code{\link{print.summary.moretrees_compact}},
  \code{\link{print.summary.moretrees_long}}
}
\concept{MOReTrees results}
