% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hyperparams_logistic_moretrees.R
\name{update_hyperparams_logistic_moretrees}
\alias{update_hyperparams_logistic_moretrees}
\title{\code{update_hyperparams_logistic_moretrees} Performs hyperparameter updates and computes 
  current value of ELBO in VI algorithm for bernoulli outcomes.}
\usage{
update_hyperparams_logistic_moretrees(X, W, y, outcomes_units, ancestors,
  levels, n, K, p, m, Fg, prob, mu, Sigma, Sigma_det, tau_t, delta, Omega,
  Omega_det, a, b, a_t_tau, b_t_tau, a_t_omega, b_t_omega, eta, g_eta,
  a_tau, b_tau, a_omega, b_omega)
}
\description{
\code{update_hyperparams_logistic_moretrees} Performs hyperparameter updates and computes 
  current value of ELBO in VI algorithm for bernoulli outcomes.
}
