% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moretrees.R, R/moretrees_wrapper.R
\docType{package}
\name{moretrees}
\alias{moretrees}
\alias{moretrees-package}
\title{moretrees: A package for fitting Multi-Outcome Regression with 
Tree-structured Shrinkage (MOReTreeS)}
\usage{
moretrees(Xcase, Xcontrol, Wcase = NULL, Wcontrol = NULL, outcomes, tr,
  vi_params_init = list(), hyperparams_init = list(),
  ci_level = 0.95, get_ml = TRUE, update_hyper_freq = 50,
  print_freq = 50, hyper_fixed = NULL, tol = 1e-08,
  tol_hyper = 1e-04, max_iter = 5000, nrestarts = 3,
  keep_restarts = TRUE, parallel = TRUE, log_restarts = FALSE,
  log_dir = ".", random_init = FALSE,
  random_init_vals = list(omega_lims = c(0.5, 1.5), tau_lims = c(0.5,
  1.5), eta_sd_frac = 0.2, mu_sd_frac = 0.2, delta_sd_frac = 0.2, u_sd_frac
  = 0.2))
}
\arguments{
\item{Xcase}{An n x K matrix of exposure data for cases, where K is the dimension of the exposure.
Grouping of the outcomes is based on their associations with variables in Xcase.
Rows of Xcase correspond to inividual cases, columns correspond to variables.}

\item{Xcontrol}{An n x K matrix of exposure data for controls; row i in Xcontrol is the matched
control for case i.}

\item{Wcase}{An n x m matrix of covariate data for cases, where m is the dimension of the exposure.
Coefficients for these variables do not affect grouping of the outcomes.
Rows of Wcase correspond to inividual cases, columns correspond to variables.}

\item{Wcontrol}{An n x m matrix of covariate data for controls; row i in Wcontrol is the matched
control for case i.}

\item{outcomes}{Character vector of length n. outcomes[i] is a string indicating the 
outcome experienced by unit i.}

\item{tr}{A directed igraph object. This is a tree representing the relationships
among the outcomes. The leaves represent individual outcomes, and internal nodes
represent outcome categories consisting of their leaf descendants. All nodes
of tr must have unique names as given by names(V(tr)). The names of the leaves must 
be equal to the unique elements of outcomes. The vertices of tr, V(tr), may have 
an attribute 'levels' containing integer values from 1 to max(V(tr)$levels). 
In this case, the levels attribute specifies groups of nodes that share common 
hyperparameters rho[f], tau[f], and omega[f]. If V(tr)$levels is NULL, 
the default is two levels of hyperparameters: one for all leaf nodes, and one
for all internal nodes.}

\item{ci_level}{A number between 0 and 1 giving the desired credible interval. 
For example, ci_level = 0.95 (the default) returns a 95\% credible interval}

\item{get_ml}{If TRUE, moretrees will also return the maximum likelihood estimates of the
coefficients for each outcome group discovered by the model. 
Default is TRUE.}

\item{update_hyper_freq}{How frequently to update hyperparameters. 
Default = every 50 iterations.}

\item{print_freq}{How often to print out iteration number and current value of epsilon
(the difference in objective function value for the two most recent iterations).}

\item{hyper_fixed}{Fixed values of hyperprior parameters for rho.
This should be a list with two elements: 
a and b, both numeric vectors of length L, representing the 
parameters of the beta prior on rho for each level, where L is the
number of levels.
Default is list(a = rep(1, L), b = rep(1, L)).}

\item{tol}{Convergence tolerance for ELBO. 
Default is 1E-8.}

\item{tol_hyper}{If hyper_method = "EB", the convergence tolerance for ELBO
between subsequent hyperparmeter updates. Typically a more generous
tolerance than tol. 
Default is 1E-4.}

\item{nrestarts}{Number of random re-starts of the VI algorithm. The result that 
gives the highest ELBO will be returned. It is recommended to choose nrestarts > 1.
The default is 3.}

\item{keep_restarts}{If TRUE, the results from all random restarts will be returned.
If FALSE, only the restart with the highest ELBO is returned. Default is TRUE.}

\item{parallel}{If TRUE, the random restarts will be run in parallel.
It is recommended to first set the number of cores using doParallel::registerDoParallel(). 
Otherwise, the default number of cores specified by the doParallel package will be used.
Default is TRUE.}

\item{log_restarts}{If TRUE, when nrestarts > 1 progress of each random restart will be 
logged to a text file in log_dir. If FALSE and nrestarts > 1, progress will not be shown.
If nrestarts = 1, progress will always be printed to the console.
Default is FALSE.}

\item{log_dir}{Directory for logging progress of random restarts.
Default is the working directory.}

\item{random_init}{The initial values for the MOReTreeS model are selected based
on maximum likelihood estimates of the effect size at every level of the tree.
If random_init = TRUE, some randomness will be added to these initial values.
The default is FALSE, unless nrestarts > 1, in which case random_init will be set
to TRUE and a warning message will be printed.}

\item{random_init_vals}{If random_init = TRUE, 
this is a list containing the following parameters for randomly permuting 
the inital values:
1. tau_lims: a vector of length 2, where tau_lims[1] is between 0 and 1,
and tau_lims[2] > 1. The initial values for the hyperparameter tau will
be chosen uniformly at random in the range (tau_init \* tau_lims[1], tau_init \* tau_lims[2]),
where tau_init is the initial value for tau either supplied in initial_values or guessed
using moretrees_init_logistic().
1. omega_lims: a vector of length 2, where omega_lims[1] is between 0 and 1,
and omega_lims[2] > 1. The initial values for the hyperparameter omega will
be chosen uniformly at random in the range (omega_init \* omega_lims[1], omega_init \* omega_lims[2]),
where omega_init is the initial value for omega either supplied in initial_values or guessed
using moretrees_init_logistic().
1. eta_sd_frac: a value between 0 and 1. The initial values for the auxilliary parameters
eta will have a normal random variate added to them with standard deviation equal to 
eta_sd_frac multiplied by the initial value for eta either supplied in initial_values or guessed
using moretrees_init_logistic(). Absolute values are then taken for any 
values of eta that are < 0.
1. mu_sd_frac: a value between 0 and 1. The initial values for the exposure effects
mu will have a normal random variate added to them with standard deviation equal to 
mu_sd_frac multiplied by the absolute value of the initial value for mu either supplied in 
initial_values or guessed using moretrees_init_logistic().
1. delta_sd_frac: a value between 0 and 1. The initial values for the exposure effects
delta will have a normal random variate added to them with standard deviation equal to 
delta_sd_frac multiplied by the absolute value of the initial value for delta either supplied in 
initial_values or guessed using moretrees_init_logistic().
1. u_sd_frac: a value between 0 and 1. The initial value for the node inclusion probabilities
will first be transformed to the log odds scale to obtain u. A normal random variate will be
added to u with standard deviation eqaul to u_sd_frac multiplied by the absolute value of the
initial value for u either supplied by the user or guessed using moretrees_init_logistic().
u will then be transformed back to the probability scale.}

\item{maxiter}{Maximum number of iterations of the VI algorithm.
Default is 5000.}
}
\value{
A list containing the following elements:
1. estimated coefficients and credible intervals; 
2. outputs from variational inference algorithm
}
\description{
Fit MOReTreeS to: normally distributed outcome data (moretrees_normal) or
binary data (moretrees_logistic).

All the details go here!
}
\section{Here are some functions}{

All about functions!
}

\section{Model Description}{

Describe MOReTreeS model and all parameters here.
}

\seealso{
Other MOReTreeS functions: \code{\link{moretrees_compute_betas}},
  \code{\link{moretrees_compute_thetas}},
  \code{\link{moretrees_design_tree}},
  \code{\link{moretrees_init_logistic}}
}
\concept{MOReTreeS functions}
