% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moretrees_compute_betas.R
\name{moretrees_compute_betas}
\alias{moretrees_compute_betas}
\title{Here's a brief description.
\code{moretrees_compute_betas} Computes MOReTreeS estimates and credible
  intervals for exposure effects from model output}
\usage{
moretrees_compute_betas(mod, ci_level, A_leaves)
}
\arguments{
\item{mod}{List containing outputs from spike and slab VI algorithm}

\item{ci_level}{A number between 0 and 1 giving the desired credible interval.
For example, ci_level = 0.95 (the default) returns a 95% credible interval.}

\item{A_leaves}{pL x p sparse ancestor Matrix where rows correspond to leaves
of tree (outcomes) and columns correspond to nodes on tree. Results in mod
must have same ordering as columns of A_leaves.}
}
\value{
A list containing the following elements:
1. beta_moretrees: estimated coefficients and credible intervals by group;
2. beta_est: estimated coefficients and credible intervals by outcome;
}
\description{
All the details go here!
}
\section{Details}{
NA
}

\seealso{
Other spike and slab functions: \code{\link{compute_betas}},
  \code{\link{compute_thetas}}, \code{\link{ml_by_group}},
  \code{\link{moretrees_compute_thetas}},
  \code{\link{moretrees_design_matrix}},
  \code{\link{moretrees}},
  \code{\link{spike_and_slab_logistic}},
  \code{\link{spike_and_slab_normal}},
  \code{\link{spike_and_slab}}
}
\concept{spike and slab functions}
