% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moretrees_init_logistic.R
\name{moretrees_init_logistic}
\alias{moretrees_init_logistic}
\title{Here's a brief description.
  \code{moretrees_init_logistic} Generates starting values for moretrees
  models.}
\usage{
moretrees_init_logistic(X, W, y, A, outcomes_units, outcomes_nodes,
  ancestors, levels, xxT, wwT, vi_params, hyperparams, hyper_fixed,
  random_init, random_init_vals)
}
\arguments{
\item{vi_params_init}{A list with any starting values supplied by the user for
VI parameters. NULL values will be filled in.}

\item{hyperparams_init}{A list with any starting values supplied by the user for
hyperparameters. NULL values will be filled in.}
}
\value{
A list containing starting values for both VI and hyper parameters
}
\description{
Here's a brief description.
  \code{moretrees_init_logistic} Generates starting values for moretrees
  models.
}
\seealso{
Other Internal VI functions: \code{\link{spike_and_slab_logistic_moretrees}},
  \code{\link{update_hyperparams_logistic_moretrees}},
  \code{\link{update_vi_params_logistic_moretrees}}
}
\concept{Internal VI functions}
