% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moretrees_init_logistic.R
\name{moretrees_init_logistic}
\alias{moretrees_init_logistic}
\title{Here's a brief description.
  \code{moretrees_init_rand} Randomly generates starting values for moretrees
  models. Not recommended if the model is converging slowly!}
\usage{
moretrees_init_logistic(X, W, y, A, outcomes_units, outcomes_nodes,
  ancestors, levels, xxT, wwT, vi_params, hyperparams, hyper_fixed,
  random_init, random_init_vals)
}
\arguments{
\item{vi_params_init}{A list with any starting values supplied by the user for
VI parameters. NULL values will be filled in.}

\item{hyperparams_init}{A list with any starting values supplied by the user for
hyperparameters. NULL values will be filled in.}
}
\value{
A list containing starting values for both VI and hyper parameters
}
\description{
Here's a brief description.
  \code{moretrees_init_rand} Randomly generates starting values for moretrees
  models. Not recommended if the model is converging slowly!
}
\section{Model Description}{

Describe MOReTreeS model and all parameters here.
}

\seealso{
Other MOReTreeS functions: \code{\link{moretrees_compute_betas}},
  \code{\link{moretrees_compute_thetas}},
  \code{\link{moretrees_design_tree}},
  \code{\link{moretrees}}
}
\concept{MOReTreeS functions}
