% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spike_and_slab_logistic_moretrees.R
\name{spike_and_slab_logistic_moretrees}
\alias{spike_and_slab_logistic_moretrees}
\title{Group spike and slab variable selection with Gaussian outcome}
\usage{
spike_and_slab_logistic_moretrees(dsgn, initial_values, vi_params_init,
  hyperparams_init, random_init, random_init_vals, tol, tol_hyper,
  max_iter, print_freq, update_hyper_freq, hyper_fixed)
}
\arguments{
\item{dsgn}{A list containing all data elements needed for the algorithm.}

\item{random_init}{If TRUE, initial values will be randomly permuted.}

\item{random_init_vals}{If random_init = TRUE, 
this is a list containing parameters for randomly permuting the inital values.
The list contains the following elements:}

\item{tol}{Convergence tolerance for ELBO.}

\item{print_freq}{How often to print out iteration number.}

\item{update_hyper_freq}{How frequently to update hyperparameters.}

\item{hyper_fixed}{Fixed values of hyperprior parameters.}

\item{maxiter}{Maximum number of iterations of the VI algorithm.}
}
\value{
A list of variational parameters.
}
\description{
Here's a brief description.
  \code{spike_and_slab_logistic_moretrees} performs group variable selection via a spike
  and slab prior for binary data.
  The posterior is approximated via variational inference.
  This function returns the parameters of the variational approximation.
}
\details{
All the details go here!
}
\section{Model Description}{

  Describe group spike and slab prior and all parameters here.
}

\seealso{
Other spike and slab functions: \code{\link{ml_by_group}}
}
\concept{spike and slab functions}
