% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_hyperparams_logistic_moretrees.R
\name{update_hyperparams_logistic_moretrees}
\alias{update_hyperparams_logistic_moretrees}
\title{\code{update_hyperparams_logistic_moretrees} Performs hyperparameter updates and computes 
  current value of ELBO in VI algorithm for bernoulli outcomes.}
\usage{
update_hyperparams_logistic_moretrees(X, W, y, outcomes_units, ancestors,
  levels, n, K, p, m, Fg, prob, mu, Sigma, Sigma_det, tau_t, delta, Omega,
  Omega_det, a_t, b_t, eta, g_eta, tau, omega, a, b, update_hyper)
}
\description{
\code{update_hyperparams_logistic_moretrees} Performs hyperparameter updates and computes 
  current value of ELBO in VI algorithm for bernoulli outcomes.
}
