% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_vi_params_logistic_moretrees.R
\name{update_vi_params_logistic_moretrees}
\alias{update_vi_params_logistic_moretrees}
\title{\code{update_vi_logistic_moretrees} Performs variational updates in VI algorithm.}
\usage{
update_vi_params_logistic_moretrees(X, W, y, xxT, wwT, outcomes_nodes,
  outcomes_units, ancestors, levels, n, p, pL, K, m, Fg, prob, mu, Sigma,
  Sigma_inv, Sigma_det, tau_t, delta, Omega, Omega_inv, Omega_det, a_t,
  b_t, eta, g_eta, tau, omega, a, b)
}
\description{
\code{update_vi_logistic_moretrees} Performs variational updates in VI algorithm.
}
\seealso{
Other Internal VI functions: \code{\link{moretrees_init_logistic}},
  \code{\link{spike_and_slab_logistic_moretrees}},
  \code{\link{update_hyperparams_logistic_moretrees}}
}
\concept{Internal VI functions}
