% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_vi_params_logistic_moretrees.R
\name{update_vi_params_logistic_moretrees}
\alias{update_vi_params_logistic_moretrees}
\title{\code{update_vi_logistic_moretrees} Performs variational updates in VI algorithm.}
\usage{
update_vi_params_logistic_moretrees(X, W, y, outcomes_nodes,
  outcomes_units, ancestors, levels, xxT, wwT, n, p, pL, K, m, Fg, prob,
  mu, Sigma, Sigma_inv, Sigma_det, tau_t, delta, Omega, Omega_inv,
  Omega_det, a_t, b_t, eta, g_eta, tau, omega, a, b)
}
\arguments{
\item{X, W, y, outcomes_units, outcomes_nodes, ancestors, levels}{outputs from 
\code{moretrees_design_tree}}

\item{xxT, wwT}{computed from \code{X} and \code{W} in \code{spike_and_slab_logisitic_moretrees()}}

\item{n, K, p, pL, m, Fg}{computed from the data by \code{spike_and_slab_logistic_moretrees()}}

\item{prob, mu, Sigma, Sigma_inv, Sigma_det, tau_t, delta, Omega, Omega_inv, Omega_det, a_t, b_t}{variational parameters updated by \code{update_vi_params_logistic_moretrees()}}

\item{eta, g_eta, tau, omega}{parameters updated by \code{update_hyperparams_logistic_moretrees()}}

\item{a, b}{fixed hyperparameters}
}
\description{
\code{update_vi_logistic_moretrees} Performs variational updates in VI algorithm.
}
\seealso{
Other Internal VI functions: \code{\link{moretrees_init_logistic}},
  \code{\link{spike_and_slab_logistic_moretrees}},
  \code{\link{update_hyperparams_logistic_moretrees}}
}
\concept{Internal VI functions}
