% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_vi_params_logistic_moretrees.R
\name{update_vi_params_logistic_moretrees}
\alias{update_vi_params_logistic_moretrees}
\title{\code{update_vi_logistic_moretrees} Performs variational updates for bernoulli outcomes.}
\usage{
update_vi_params_logistic_moretrees(X, W, y, xxT, wwT, outcomes_nodes,
  outcomes_units, ancestors, levels, n, p, pL, K, m, Fg, prob, mu, Sigma,
  Sigma_inv, Sigma_det, tau_t, delta, Omega, Omega_inv, Omega_det, a_t,
  b_t, eta, g_eta, tau, omega, a, b)
}
\description{
\code{update_vi_logistic_moretrees} Performs variational updates for bernoulli outcomes.
}
